Use Array.reduce to sort even/odd numbers
Call Array.reduce on theArray passing in an initial value of an array
with 2 empty arrays as entries ( [[], []] ).  Array.reduce will call
the callback for each entry in theArray.  If the value is even, put it
in the first array.  If the value is odd, put it in the second array.
When Array.reduce is complete, concatonate first array on to the second
array.

Array.reduce is available in all modern browsers:
https://developer.mozilla.org/en-US/docs/JavaScript/Reference/Global_Obj
ects/Array/Reduce#Browser_compatibility